+++ baybe/kernels/base.py
@@ -3,21 +3,24 @@
 from __future__ import annotations
 
 import gc
-from abc import ABC
-from collections.abc import Sequence
+from abc import ABC, abstractmethod
+from itertools import chain
 from typing import TYPE_CHECKING, Any
 
-from attrs import define
+from attrs import define, field
+from attrs.converters import optional as optional_c
+from attrs.validators import deep_iterable, instance_of
+from attrs.validators import optional as optional_v
+from typing_extensions import override
 
 from baybe.exceptions import UnmatchedAttributeError
 from baybe.priors.base import Prior
+from baybe.searchspace.core import SearchSpace
 from baybe.serialization.mixin import SerialMixin
 from baybe.settings import active_settings
-from baybe.utils.basic import get_baseclasses, match_attributes
+from baybe.utils.basic import classproperty, get_baseclasses, match_attributes, to_tuple
 
 if TYPE_CHECKING:
-    import torch
-
     from baybe.surrogates.gaussian_process.components.kernel import PlainKernelFactory
 
 
@@ -83,6 +86,11 @@ class Kernel(ABC, SerialMixin):
         # Enable use with math.prod(), which starts with 1 * first_element.
         return self.__mul__(other)
 
+    @classproperty
+    def _whitelisted_attributes(cls) -> frozenset[str]:
+        """Attribute names to exclude from gpytorch matching."""
+        return frozenset()
+
     def to_factory(self) -> PlainKernelFactory:
         """Wrap the kernel in a :class:`baybe.surrogates.gaussian_process.components.PlainKernelFactory`."""  # noqa: E501
         from baybe.surrogates.gaussian_process.components.kernel import (
@@ -91,24 +99,17 @@ class Kernel(ABC, SerialMixin):
 
         return PlainKernelFactory(self)
 
-    def to_gpytorch(
-        self,
-        *,
-        ard_num_dims: int | None = None,
-        batch_shape: torch.Size | None = None,
-        active_dims: Sequence[int] | None = None,
-    ):
+    @abstractmethod
+    def _get_dimensions(
+        self, searchspace: SearchSpace
+    ) -> tuple[tuple[int, ...] | None, int | None]:
+        """Get the active dimensions and the number of ARD dimensions."""
+
+    def to_gpytorch(self, searchspace: SearchSpace):
         """Create the gpytorch representation of the kernel."""
         import gpytorch.kernels
 
-        # Extract keywords with non-default values. This is required since gpytorch
-        # makes use of kwargs, i.e. differentiates if certain keywords are explicitly
-        # passed or not. For instance, `ard_num_dims = kwargs.get("ard_num_dims", 1)`
-        # fails if we explicitly pass `ard_num_dims=None`.
-        kw: dict[str, Any] = dict(
-            ard_num_dims=ard_num_dims, batch_shape=batch_shape, active_dims=active_dims
-        )
-        kw = {k: v for k, v in kw.items() if v is not None}
+        active_dims, ard_num_dims = self._get_dimensions(searchspace)
 
         # Get corresponding gpytorch kernel class and its base classes
         try:
@@ -134,10 +135,12 @@ class Kernel(ABC, SerialMixin):
             unmatched_attrs.update(unmatched)
 
         # Sanity check: all attributes of the BayBE kernel need a corresponding match
-        # in the gpytorch kernel (otherwise, the BayBE kernel class is misconfigured).
-        # Exceptions: initial values and trainability flags are not used during
-        # construction but are set on the created object after construction.
-        missing = set(unmatched) - set(kernel_attrs)
+        # with the gpytorch kernel signature (otherwise, the BayBE kernel class is
+        # misconfigured). Exceptions: initial values and trainability flags are not used
+        # during construction but are set on the created object after construction.
+        missing = (
+            set(unmatched_attrs) - set(kernel_attrs) - self._whitelisted_attributes
+        )
         if leftover := {
             m
             for m in missing
@@ -154,7 +157,7 @@ class Kernel(ABC, SerialMixin):
 
         # Convert specified inner kernels to gpytorch, if provided
         kernel_dict = {
-            key: value.to_gpytorch(**kw)
+            key: value.to_gpytorch(searchspace)
             for key, value in kernel_attrs.items()
             if isinstance(value, Kernel)
         }
@@ -162,7 +165,9 @@ class Kernel(ABC, SerialMixin):
         # Create the kernel with all its inner gpytorch objects
         kernel_attrs.update(kernel_dict)
         kernel_attrs.update(prior_dict)
-        gpytorch_kernel = kernel_cls(**kernel_attrs, **kw)
+        gpytorch_kernel = kernel_cls(
+            **kernel_attrs, ard_num_dims=ard_num_dims, active_dims=active_dims
+        )
 
         # If the kernel has a lengthscale, set its initial value
         if kernel_cls.has_lengthscale:
@@ -185,10 +190,57 @@ class Kernel(ABC, SerialMixin):
 class BasicKernel(Kernel, ABC):
     """Abstract base class for all basic kernels."""
 
+    parameter_names: tuple[str, ...] | None = field(
+        default=None,
+        converter=optional_c(to_tuple),
+        validator=optional_v(
+            deep_iterable(
+                iterable_validator=instance_of(tuple),
+                member_validator=instance_of(str),
+            )
+        ),
+        kw_only=True,
+    )
+    """An optional set of names specifiying the parameters the kernel should act on."""
+
+    @override
+    @classproperty
+    def _whitelisted_attributes(cls) -> frozenset[str]:
+        return frozenset({"parameter_names"})
+
+    @override
+    def _get_dimensions(
+        self, searchspace: SearchSpace
+    ) -> tuple[tuple[int, ...] | None, int | None]:
+        if self.parameter_names is None:
+            # `None` is gpytorch's default indicating that all dimensions are active
+            active_dims = None
+        else:
+            active_dims = tuple(
+                chain.from_iterable(
+                    searchspace.get_comp_rep_parameter_indices(name)
+                    for name in self.parameter_names
+                )
+            )
+
+        # We use automatic relevance determination for all kernels
+        ard_num_dims = (
+            len(active_dims)
+            if active_dims is not None
+            else len(searchspace.comp_rep_columns)
+        )
+        return active_dims, ard_num_dims
+
 
 @define(frozen=True)
 class CompositeKernel(Kernel, ABC):
     """Abstract base class for all composite kernels."""
+
+    @override
+    def _get_dimensions(
+        self, searchspace: SearchSpace
+    ) -> tuple[tuple[int, ...] | None, int | None]:
+        return None, None
 
 
 # Collect leftover original slotted classes processed by `attrs.define`