+++ baybe/surrogates/gaussian_process/presets/edbo.py
@@ -4,20 +4,28 @@ from __future__ import annotations
 
 import gc
 from collections.abc import Collection
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, ClassVar
 
-from attrs import define
+from attrs import define, field
 from typing_extensions import override
 
 from baybe.kernels.basic import MaternKernel
 from baybe.kernels.composite import ScaleKernel
 from baybe.parameters import TaskParameter
 from baybe.parameters.enum import SubstanceEncoding
+from baybe.parameters.selectors import (
+    ParameterSelectorProtocol,
+    TypeSelector,
+    _ParameterSelectorMixin,
+    to_parameter_selector,
+)
 from baybe.parameters.substance import SubstanceParameter
 from baybe.priors.basic import GammaPrior
 from baybe.searchspace.discrete import SubspaceDiscrete
-from baybe.surrogates.gaussian_process.components.kernel import KernelFactory
-from baybe.surrogates.gaussian_process.components.likelihood import LikelihoodFactory
+from baybe.surrogates.gaussian_process.components.kernel import KernelFactoryProtocol
+from baybe.surrogates.gaussian_process.components.likelihood import (
+    LikelihoodFactoryProtocol,
+)
 from baybe.surrogates.gaussian_process.components.mean import LazyConstantMeanFactory
 
 if TYPE_CHECKING:
@@ -48,13 +56,22 @@ _EDBO_ENCODINGS = (
 
 
 @define
-class EDBOKernelFactory(KernelFactory):
+class EDBOKernelFactory(KernelFactoryProtocol, _ParameterSelectorMixin):
     """A factory providing EDBO kernels.
 
     References:
         * https://github.com/b-shields/edbo/blob/master/edbo/bro.py#L664
         * https://doi.org/10.1038/s41586-021-03213-y
     """
+
+    _uses_parameter_names: ClassVar[bool] = True
+    # See base class.
+
+    parameter_selector: ParameterSelectorProtocol | None = field(
+        factory=lambda: TypeSelector([TaskParameter], exclude=True),
+        converter=to_parameter_selector,
+    )
+    # TODO: Reuse base attribute (https://github.com/python-attrs/attrs/pull/1429)
 
     @override
     def __call__(
@@ -101,6 +118,7 @@ class EDBOKernelFactory(KernelFactory):
                 nu=2.5,
                 lengthscale_prior=lengthscale_prior,
                 lengthscale_initial_value=lengthscale_initial_value,
+                parameter_names=self.get_parameter_names(searchspace),
             ),
             outputscale_prior=outputscale_prior,
             outputscale_initial_value=outputscale_initial_value,
@@ -112,7 +130,7 @@ EDBOMeanFactory = LazyConstantMeanFactory
 
 
 @define
-class EDBOLikelihoodFactory(LikelihoodFactory):
+class EDBOLikelihoodFactory(LikelihoodFactoryProtocol):
     """A factory providing EDBO likelihoods.
 
     References: